Moving updating home screen to screens folder
